requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OpenAI and AI Models
openai==1.3.8
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2

# Vector Database
chromadb==0.4.18

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0

# Data processing
pandas==2.1.4
numpy>=1.26.0,<2.0.0
nltk==3.8.1
spacy==3.7.2

# PDF Processing
PyMuPDF==1.23.14
pdfplumber==0.10.4
pypdf==3.17.4
python-docx==1.1.0

# OCR Support
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
opencv-python==4.8.1.78

# Machine Learning
scikit-learn==1.3.2

# Environment and Configuration
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# Progress bars and utilities
tqdm==4.66.1

# HTTP and async support
aiohttp==3.9.1
httpx==0.25.2
aiofiles==23.2.1

# Google Sheets integration
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
charset-normalizer==3.3.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
